fix clang-cuda

+++ libcudacxx/test/libcudacxx/cuda/annotated_ptr/memcpy_async.pass.cpp
@@ -87,7 +87,7 @@ __device__ __host__ __noinline__ void test_memcpy_async()
         if (arr1[i] != (int) i)
         {
           DPRINTF(stderr, "%p:&arr1[i] == %d, should be:%lu\n", &arr1[i], arr1[i], i);
-          assert(arr1[i] == i);
+          assert(arr1[i] == (int) i);
         }
 
         arr1[i] = 0;
@@ -101,7 +101,7 @@ __device__ __host__ __noinline__ void test_memcpy_async()
         if (arr1[i] != (int) i)
         {
           DPRINTF(stderr, "%p:&arr1[i] == %d, should be:%lu\n", &arr1[i], arr1[i], i);
-          assert(arr1[i] == i);
+          assert(arr1[i] == (int) i);
         }
 
         arr1[i] = 0;
